requirements.txt
requests
python-dotenv
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
bcrypt
python-jose[cryptography]
python-multipart
httpx
email-validator
anthropic
google-cloud-translate
google-auth-oauthlib
pyotp
